// frontend/src/api/model/bedReservationRead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */
import type { BedReservationStatus } from './bedReservationStatus';

export interface BedReservationRead {
  id: string;
  hospitalization_id: string;
  bed_id: string;
  status: BedReservationStatus;
  reserved_at: string;
  expires_at: string | null;
  reserved_by: string | null;
  completed_at: string | null;
  cancelled_at: string | null;
  reason: string | null;
}